requirements.txt
discord.py>=2.3.0
motor>=3.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
dnspython>=2.3.0
pydantic_settings>=2.12.0
chat-exporter==3.0.1
